+++ Riot/Modules/CreateRoom/EnterNewRoomDetails/EnterNewRoomDetailsViewController.swift
@@ -254,7 +254,7 @@ final class EnterNewRoomDetailsViewController: UIViewController {
         mainTableView.register(cellType: MXKTableViewCellWithTextView.self)
         mainTableView.register(cellType: TextFieldTableViewCell.self)
         mainTableView.register(cellType: TextViewTableViewCell.self)
-        mainTableView.register(headerFooterViewType: TableViewHeaderFooterView.self)
+        mainTableView.register(headerFooterViewType: TextViewTableViewHeaderFooterView.self)
         mainTableView.sectionHeaderHeight = UITableView.automaticDimension
         mainTableView.estimatedSectionHeaderHeight = 50
         mainTableView.sectionFooterHeight = UITableView.automaticDimension
@@ -412,7 +412,7 @@ extension EnterNewRoomDetailsViewController: UITableViewDelegate {
             return nil
         }
 
-        let view: TableViewHeaderFooterView? = tableView.dequeueReusableHeaderFooterView()
+        let view: TextViewTableViewHeaderFooterView? = tableView.dequeueReusableHeaderFooterView()
 
         view?.textView.text = header
         view?.textView.font = .systemFont(ofSize: 13)
@@ -427,7 +427,7 @@ extension EnterNewRoomDetailsViewController: UITableViewDelegate {
             return nil
         }
 
-        let view: TableViewHeaderFooterView? = tableView.dequeueReusableHeaderFooterView()
+        let view: TextViewTableViewHeaderFooterView? = tableView.dequeueReusableHeaderFooterView()
 
         view?.textView.text = footer
         view?.textView.font = .systemFont(ofSize: 13)