+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes to be released in next version
 
 🙌 Improvements
  * Pin: Implement not allowed PINs feature. There is no restriction by default.
+ * Room: New Room Settings screen.
  * Architecture: Use coordinator pattern for legacy screen flows (#3597).
 
 🐛 Bugfix